Refactor code for improved readability and maintainability
Refactored multiple command classes to adhere to good coding practices.
Changes include:
- Improved variable naming for better clarity.
- Added meaningful comments where necessary and removed redundant ones.
- Ensured immutability by marking fields as `final` where appropriate.
- Extracted complex logic into smaller methods to follow the
Single Responsibility Principle.
- Enhanced error handling with more descriptive exceptions.

This refactor improves the overall maintainability and clarity of the
code, following the SE-EDU Java coding standards.
