--- src/main/java/GUI/Launcher.java
@@ -1,13 +1,0 @@
-package GUI;
-
-import javafx.application.Application;
-
-/**
- * A launcher class to workaround classpath issues.
- */
-public class Launcher {
-    public static void main(String[] args) {
-        Application.launch(Main.class, args);
-    }
-}
-
