--- src/main/java/GUI/Main.java
@@ -1,36 +1,0 @@
-package GUI;
-
-import java.io.IOException;
-
-import javafx.application.Application;
-import javafx.fxml.FXMLLoader;
-import javafx.scene.Scene;
-import javafx.scene.layout.AnchorPane;
-import javafx.stage.Stage;
-
-import Main.Diego;
-
-
-
-/**
- * A GUI for Diego using FXML.
- */
-public class Main extends Application {
-
-    private Diego diego = new Diego("data/Diego.txt");
-
-    @Override
-    public void start(Stage stage) {
-
-        try {
-            FXMLLoader fxmlLoader = new FXMLLoader(Main.class.getResource("/view/MainWindow.fxml"));
-            AnchorPane ap = fxmlLoader.load();
-            Scene scene = new Scene(ap);
-            stage.setScene(scene);
-            fxmlLoader.<MainWindow>getController().setDiego(diego); // inject the Diego instance
-            stage.show();
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
-    }
-}
