--- src/main/java/Tasks/Event.java
@@ -1,51 +1,0 @@
-package Tasks;
-
-/**
- * Represents an event task with a description, start time, and end time.
- */
-public class Event extends Task {
-    protected String start;
-    protected String end;
-
-    /**
-     * Constructs a new Event task.
-     *
-     * @param description The description of the event task.
-     * @param start       The start time/date of the event.
-     * @param end         The end time/date of the event.
-     */
-    public Event(String description, String start, String end) {
-        super(description);
-        this.start = start;
-        this.end = end;
-    }
-
-    /**
-     * Returns a string representation of the event task.
-     *
-     * @return A string describing the event task.
-     */
-    @Override
-    public String toString() {
-        return "[E]" + (isDone ? "[X]" : "[ ]") + " " + description + " (from: " + start + " to: " + end + ")";
-    }
-
-    /**
-     * Returns the file format string of the event task for storage.
-     *
-     * @return A string suitable for storing in a file.
-     */
-    @Override
-    public String toFileFormat() {
-        return "E | " + (isDone ? "1" : "0") + " | " + description + " | " + start + " | " + end;
-    }
-
-    /**
-     * Checks if the task is due within the next specified number of days.
-     * This is a default implementation that does nothing for tasks without deadlines/events.
-     *
-     * @param days The number of days to check for upcoming deadlines.
-     * @return false by default for tasks with no due date.
-     */
-
-}
