+++ src/main/java/gui/DialogBox.java
@@ -1,0 +1,66 @@
+package gui;
+
+import java.io.IOException;
+import java.util.Collections;
+
+import javafx.collections.FXCollections;
+import javafx.collections.ObservableList;
+import javafx.fxml.FXML;
+import javafx.fxml.FXMLLoader;
+import javafx.geometry.Pos;
+import javafx.scene.Node;
+import javafx.scene.control.Label;
+import javafx.scene.image.Image;
+import javafx.scene.image.ImageView;
+import javafx.scene.layout.HBox;
+
+/**
+ * Represents a dialog box consisting of an ImageView to represent the speaker's face
+ * and a label containing text from the speaker.
+ */
+public class DialogBox extends HBox {
+    @FXML
+    private Label dialog;
+    @FXML
+    private ImageView displayPicture;
+
+    private DialogBox(String text, Image img) {
+        assert text != null && !text.trim().isEmpty() : "Dialog text should not be null or empty";
+        assert img != null : "Image should not be null";
+
+        loadFxml();
+        dialog.setText(text);
+        displayPicture.setImage(img);
+    }
+
+    private void loadFxml() {
+        try {
+            FXMLLoader fxmlLoader = new FXMLLoader(MainWindow.class.getResource("/view/DialogBox.fxml"));
+            fxmlLoader.setController(this);
+            fxmlLoader.setRoot(this);
+            fxmlLoader.load();
+        } catch (IOException e) {
+            e.printStackTrace();
+        }
+    }
+
+    /**
+     * Flips the dialog box so the ImageView is on the left and text on the right.
+     */
+    private void flip() {
+        ObservableList<Node> nodes = FXCollections.observableArrayList(this.getChildren());
+        Collections.reverse(nodes);
+        getChildren().setAll(nodes);
+        setAlignment(Pos.TOP_LEFT);
+    }
+
+    public static DialogBox getUserDialog(String text, Image img) {
+        return new DialogBox(text, img);
+    }
+
+    public static DialogBox getDiegoDialog(String text, Image img) {
+        var dialogBox = new DialogBox(text, img);
+        dialogBox.flip();
+        return dialogBox;
+    }
+}
