+++ src/main/java/gui/Launcher.java
@@ -1,0 +1,12 @@
+package gui;
+
+import javafx.application.Application;
+
+/**
+ * Launcher class to workaround classpath issues.
+ */
+public class Launcher {
+    public static void main(String[] args) {
+        Application.launch(Main.class, args);
+    }
+}
