+++ src/main/java/utils/Storage.java
@@ -1,0 +1,105 @@
+package utils;
+
+import java.io.File;
+import java.io.FileNotFoundException;
+import java.io.FileWriter;
+import java.io.IOException;
+import java.util.ArrayList;
+import java.util.Scanner;
+
+import tasks.Deadline;
+import tasks.Event;
+import tasks.Task;
+import tasks.TaskList;
+import tasks.Todo;
+
+/**
+ * Handles loading and saving tasks to and from a file.
+ */
+public class Storage {
+    private final String filePath;
+
+    /**
+     * Constructs a new Storage object.
+     *
+     * @param filePath File path where tasks are stored.
+     */
+    public Storage(String filePath) {
+        this.filePath = filePath;
+    }
+
+    /**
+     * Creates the file if it does not exist.
+     */
+    public void createFileIfNotExists() {
+        File file = new File(filePath);
+        try {
+            if (!file.exists()) {
+                file.getParentFile().mkdirs();
+                file.createNewFile();
+            }
+        } catch (IOException e) {
+            e.printStackTrace();
+        }
+    }
+
+    /**
+     * Loads tasks from the file.
+     *
+     * @return List of tasks loaded from the file.
+     * @throws FileNotFoundException If the file cannot be found.
+     */
+    public ArrayList<Task> load() throws FileNotFoundException {
+        ArrayList<Task> tasks = new ArrayList<>();
+        File file = new File(filePath);
+        try (Scanner scanner = new Scanner(file)) {
+            while (scanner.hasNextLine()) {
+                String line = scanner.nextLine();
+                Task task = parseTask(line);
+                if (task != null) {
+                    tasks.add(task);
+                }
+            }
+        }
+        return tasks;
+    }
+
+    private Task parseTask(String line) {
+        String[] parts = line.split(" \\| ");
+        if (parts.length < 3) {
+            return null;
+        }
+
+        String taskType = parts[0];
+        boolean isDone = parts[1].equals("1");
+        String description = parts[2];
+
+        Task task = switch (taskType) {
+            case "T" -> new Todo(description);
+            case "D" -> new Deadline(description, parts[3]);
+            case "E" -> new Event(description, parts[3], parts[4]);
+            default -> null;
+        };
+
+        if (task != null && isDone) {
+            task.mark();
+        }
+
+        return task;
+    }
+
+    /**
+     * Saves the task list to the file.
+     *
+     * @param tasks TaskList to be saved.
+     */
+    public void save(TaskList tasks) {
+        try (FileWriter writer = new FileWriter(filePath)) {
+            for (Task task : tasks.getAll()) {
+                writer.write(task.toFileFormat() + System.lineSeparator());
+            }
+        } catch (IOException e) {
+            e.printStackTrace();
+        }
+    }
+}
