+++ src/test/java/utils/DeadlineTest.java
@@ -1,0 +1,38 @@
+package utils;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+import org.junit.jupiter.api.Test;
+
+import tasks.Deadline;
+
+class DeadlineTest {
+
+    @Test
+    void testToString_withTime() {
+        Deadline deadline = new Deadline("submit report", "2023-09-30 2359");
+        String expected = "[D][ ] submit report (by: Sep 30 2023, 11:59 pm)";
+        assertEquals(expected, deadline.toString());
+    }
+
+    @Test
+    void testToString_withoutTime() {
+        Deadline deadline = new Deadline("submit report", "2023-09-30");
+        String expected = "[D][ ] submit report (by: Sep 30 2023)";
+        assertEquals(expected, deadline.toString());
+    }
+
+    @Test
+    void testToFileFormat_withTime() {
+        Deadline deadline = new Deadline("submit report", "2023-09-30 2359");
+        String expected = "D | 0 | submit report | 2023-09-30 2359";
+        assertEquals(expected, deadline.toFileFormat());
+    }
+
+    @Test
+    void testToFileFormat_withoutTime() {
+        Deadline deadline = new Deadline("submit report", "2023-09-30");
+        String expected = "D | 0 | submit report | 2023-09-30";
+        assertEquals(expected, deadline.toFileFormat());
+    }
+}
